add seta + verificaçoes de acesso
