+++ App.js
@@ -1,13 +1,13 @@
 import React from 'react';
 import { View, StyleSheet, StatusBar } from 'react-native';
 import Introduction_Screen from './src/screens/Introduction_Screen/Introduction_Screen';
-import TelaLogin from './src/screens/TelaLogin';
+import Login_Screen from './src/screens/Login_Screen/Login_Screen.js';
 
 export default function App() {
   return (
     <View style={styles.container}>
       <StatusBar backgroundColor="#4A4DA1"/>
-      <TelaLogin />
+      <Login_Screen />
     </View>
   );
 }
